+++ packages/terra-i18n/tests/wdio/i18n-spec.js
@@ -4,17 +4,17 @@ const testLocales = ['en', 'en-GB', 'pt', 'pt-GW', 'zu', 'zu-ZA'];
 Terra.describeViewports('I18n', ['medium'], () => {
   it('should display tests initial load', () => {
     browser.url('/raw/tests/terra-i-18-n/i-18-n/default-i-18-n');
-    browser.waitForVisible('select');
+    $('select').waitForDisplayed();
 
     Terra.validates.element('initial load');
   });
 
   testLocales.forEach((locale) => {
     it(`should display message in ${locale}`, () => {
-      browser.waitForVisible('select');
-      browser.element('select').selectByValue(locale);
+      $('select').waitForDisplayed();
+      $('select').selectByAttribute('value', locale);
 
-      Terra.validates.screenshot(`${locale} locale`);
+      Terra.validates.element(`${locale} locale`);
     });
   });
 });